+++ CHANGELOG.md
@@ -33,13 +33,14 @@ If you like this project and find it useful, please consider giving it a star on
 
 ### Breaking changes
 
-- [matterbridge]: Require matterbridge v.3.10.0 with matter v.1.6.0 and matter.js v.0.17.5.
+- [matterbridge]: Require matterbridge v.3.10.0 with matter v.1.6.0.
 
 ### Changed
 
-- [package]: Bump `oxfmt` to v.0.60.0.
-- [package]: Bump `oxlint` to v.1.75.0.
+- [package]: Bump `oxfmt` to v.0.61.0.
+- [package]: Bump `oxlint` to v.1.76.0.
 - [package]: Bump `oxlint-tsgolint` to v.7.0.2001.
+- [package]: Bump `@types/node` to v.26.1.2.
 - [package]: Update agents configs.
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>